eventsのプライマリーキーをUUIDに変更

+++ database/migrations/2024_11_26_130554_create_questions_table.php
@@ -8,13 +8,15 @@ class CreateQuestionsTable extends Migration
 {
     public function up()
     {
-        Schema::create('questions', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('event_id')->constrained('events')->onDelete('cascade');
-            $table->string('text')->comment('質問文');
-            $table->enum('type', ['multiple_choice', 'text_input'])->comment('質問の種類');
-            $table->timestamps();
-        });
+        if (Schema::hasTable('events') && !Schema::hasTable('questions')) {
+            Schema::create('questions', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('event_id')->constrained('events')->onDelete('cascade');
+                $table->string('text')->comment('質問文');
+                $table->enum('type', ['multiple_choice', 'text_input'])->comment('質問の種類');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
