+++ database/migrations/2024_11_26_130630_create_choices_table.php
@@ -8,13 +8,15 @@ class CreateChoicesTable extends Migration
 {
     public function up()
     {
-        Schema::create('choices', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('question_id')->constrained('questions')->onDelete('cascade'); // 外部キー
-            $table->string('text')->comment('選択肢');
-            $table->boolean('is_correct')->default(false)->comment('正解フラグ');
-            $table->timestamps();
-        });
+        if (Schema::hasTable('questions') && !Schema::hasTable('choices')) {
+            Schema::create('choices', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('question_id')->constrained('questions')->onDelete('cascade'); // 外部キー
+                $table->string('text')->comment('選択肢');
+                $table->boolean('is_correct')->default(false)->comment('正解フラグ');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
